Hack fix for qwt 64 bit Windows compiler warnings

+++ libs/qwt/qwt_plot_curve.cpp
@@ -349,7 +349,7 @@ void QwtPlotCurve::drawSeries( QPainter *painter,
     const QwtScaleMap &xMap, const QwtScaleMap &yMap,
     const QRectF &canvasRect, int from, int to ) const
 {
-    const size_t numSamples = dataSize();
+    const int numSamples = static_cast<int>(dataSize());
 
     if ( !painter || numSamples <= 0 )
         return;
@@ -405,7 +405,7 @@ void QwtPlotCurve::drawCurve( QPainter *painter, int style,
                 // we always need the complete
                 // curve for fitting
                 from = 0;
-                to = dataSize() - 1;
+                to = static_cast<int>(dataSize()) - 1;
             }
             drawLines( painter, xMap, yMap, canvasRect, from, to );
             break;
